DiskCollections

DiskCollections.DiskDict

DiskCollections.LoggingDict

DiskCollections.DiskDict — Method

DiskDict{K,V}(io::IO, n::Int; buffered_checkpoint=true)
DiskDict{K,V}(filename::String, n::Int; buffered_checkpoint=true)
DiskDict{K,V}(f::Function, filename::String, n::Int; buffered_checkpoint=true)

Constructs a persistent dictionary behaving like a Dict{K,V} that serializes keys and values to the seekable IO io. If a filename is provided instead that file is opened with read, write, and create set to true. Every n updates the dictionary reserializes its current state. If buffered_checkpoint is true, this reserialization is performed in memory before writing the data to the io.

Because the constructor opens a file that should be closed a function can be provided as the first argument, enabling the use of Do-Block Syntax just as for open.

julia> new_disk_dict = DiskDict(open("mydict.jlp","w+"), 1000)
DiskDict{Any,Any}()

julia> new_disk_dict["name"] = "DiskCollections"
"DiskCollections"

julia> close(new_disk_dict)

julia> plain_dict = DiskDict("mydict.jlp", 1000) do disk_dict
           Dict(disk_dict)
       end
Dict{Any,Any} with 1 entry:
  "name" => "DiskCollections"

DiskCollections.LoggingDict — Method

LoggingDict{K,V}(io::IO; until::Union{DateTime,Nothing}=nothing)
LoggingDict{K,V}(filename::String; until::Union{DateTime,Nothing}=nothing)
LoggingDict{K,V}(f::Function, filename::String; until::Union{DateTime,Nothing}=nothing)

Constructs a persistent dictionary behaving like a Dict{K,V} that serializes keys and values to the seekable IO io. If a filename is provided instead that file is opened with read, write, and create set to true. For each update the LoggingDict stores the time; by providing a DateTime for the until parameter of the LoggingDict constructor the updates are loaded from the file up until that point in time.

Because the constructor opens a file that should be closed a function can be provided as the first argument, enabling the use of Do-Block Syntax just as for open.

julia> using Dates

julia> LoggingDict{Int,String}("mylogdict.jlp") do logdict
           for i in 1:5
               logdict[i] = string(i)
           end
           sleep(10)
           for i in 6:10
               logdict[i] = string(i)
           end
       end

julia> LoggingDict{Int,String}("mylogdict.jlp", until=now()-Second(10))
LoggingDict{Int64,String} with 5 entries:
  4 => "4"
  2 => "2"
  3 => "3"
  5 => "5"
  1 => "1"